refactor(ui): move review-related components in review folder
