+++ packages/client/app/ui/review/index.js
@@ -1,2 +1,6 @@
 // eslint-disable-next-line import/prefer-default-export
 export { default as ReviewerForm } from './ReviewerForm'
+export { default as ReviewerAcceptButton } from './ReviewerAcceptButton'
+export { default as ReviewerRejectButton } from './ReviewerRejectButton'
+export { default as ReviewerSubmitButton } from './ReviewerSubmitButton'
+export { default as ReviewerEditorUploadButton } from './ReviewerEditorUploadButton'
